+++ api/auth/auth-router.ts
@@ -1,53 +1,15 @@
 import { Request, Response, NextFunction } from 'express';
-import jwt from 'jsonwebtoken';
 import bcryptjs from 'bcryptjs';
 import Users from '../users/user-model';
-import Secrets from '../../database/config/secrets';
-import Codes from '../../database/config/codes';
-import { User } from '../../types';
+import AuthMiddleware from './auth-middleware';
+import Authenticate from './authentication';
+import Codes from '../../config/codes';
 
 const router = require('express').Router();
 
-const validateUser = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<void> => {
-  // Ensure there's a body at all
-  if (!req.body) {
-    res.status(400).json(Codes.BAD_REQUEST);
-    return;
-  }
-
-  const { username, password } = req.body;
-
-  // Ensure there's both a username and password
-  if (username && password) {
-    const existingUser = await Users.getByUsername(username);
-
-    if (!existingUser) {
-      next();
-    } else {
-      res.status(200).json(Codes.ID_CONFLICT);
-    }
-  } else {
-    res.status(400).json(Codes.BAD_REQUEST);
-  }
-};
-
-const generateToken = (user: User): string => {
-  const payload = {
-    username: user.username,
-  };
-  const options = {
-    expiresIn: '1d',
-  };
-  return jwt.sign(payload, Secrets.jwtSecret, options);
-};
-
 router.post(
   '/register',
-  validateUser,
+  AuthMiddleware.validateUser,
   async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { username } = req.body;
@@ -69,7 +31,7 @@ router.post(
       try {
         const user = await Users.getByUsername(username);
         if (user && bcryptjs.compareSync(password, user.password)) {
-          const token = generateToken(user);
+          const token = Authenticate.generateToken(user);
           res.status(200).json({ token });
         } else {
           res.status(401).json(Codes.INVALID_CRED);
