+++ database/seeds/testing/01-users.ts
@@ -1,6 +1,6 @@
 import * as Knex from 'knex';
 import bcryptjs from 'bcryptjs';
-import Secrets from '../../config/secrets';
+import Secrets from '../../../config/secrets';
 
 // eslint-disable-next-line import/prefer-default-export
 export async function seed(knex: Knex): Promise<void> {
